refactor: 출력 로직이 BigDecimal을 받지 않도록 변경

+++ src/main/java/view/OutputView.java
@@ -1,6 +1,5 @@
 package view;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -37,11 +36,11 @@ public class OutputView {
         System.out.println("딜러는 " + Dealer.DRAW_THRESHOLD + "을 초과하여 카드를 더 받지 않았습니다." + LINE_SEPARATOR);
     }
 
-    public void printFinalProfit(Map<String, BigDecimal> profitByParticipant) {
+    public void printFinalProfit(Map<String, Long> profitByParticipant) {
         final String header = "## 최종 수익" + System.lineSeparator();
         String output = profitByParticipant.entrySet()
                 .stream()
-                .map(entry -> entry.getKey() + ": " + entry.getValue().longValue())
+                .map(entry -> entry.getKey() + ": " + entry.getValue())
                 .collect(Collectors.joining(
                         System.lineSeparator(),
                         header,
